requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
certifi==2025.1.31
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.1
    # via
    #   -r requirements.in
    #   requests
colorama==0.4.6
    # via
    #   -r requirements.in
    #   tqdm
idna==3.10
    # via
    #   -r requirements.in
    #   requests
numpy==2.2.5
    # via
    #   -r requirements.in
    #   pandas
pandas==2.2.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   pandas
pytz==2025.2
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
six==1.17.0
    # via
    #   -r requirements.in
    #   python-dateutil
tqdm==4.67.1
    # via -r requirements.in
tzdata==2025.2
    # via
    #   -r requirements.in
    #   pandas
urllib3==2.4.0
    # via
    #   -r requirements.in
    #   requests
